Update before unload form component … and allow it to work with multiple forms per page
